Fix int vs. ssize_t type confusion
Plus a typo in the target type of a static_cast.

+++ src/encoding/tlv.hpp
@@ -294,7 +294,7 @@ readType(InputIterator& begin, const InputIterator& end, uint32_t& type)
       return false;
     }
 
-  type = static_cast<uint64_t>(number);
+  type = static_cast<uint32_t>(number);
   return true;
 }
